Move pkg/cloudprovider/provider to k8s.io/legacy-cloud-providers
