+++ cmd/kube-controller-manager/app/BUILD
@@ -40,7 +40,6 @@ go_library(
         "//pkg/apis/scheduling/install:go_default_library",
         "//pkg/apis/settings/install:go_default_library",
         "//pkg/apis/storage/install:go_default_library",
-        "//pkg/cloudprovider/providers:go_default_library",
         "//pkg/controller:go_default_library",
         "//pkg/controller/apis/config:go_default_library",
         "//pkg/controller/bootstrap:go_default_library",
@@ -146,6 +145,7 @@ go_library(
         "//staging/src/k8s.io/component-base/version/verflag:go_default_library",
         "//staging/src/k8s.io/csi-translation-lib:go_default_library",
         "//staging/src/k8s.io/csi-translation-lib/plugins:go_default_library",
+        "//staging/src/k8s.io/legacy-cloud-providers/install:go_default_library",
         "//staging/src/k8s.io/metrics/pkg/client/clientset/versioned/typed/metrics/v1beta1:go_default_library",
         "//staging/src/k8s.io/metrics/pkg/client/custom_metrics:go_default_library",
         "//staging/src/k8s.io/metrics/pkg/client/external_metrics:go_default_library",
