+++ cmd/kubelet/app/plugins.go
@@ -51,7 +51,7 @@ import (
 	"k8s.io/kubernetes/pkg/volume/storageos"
 
 	// Cloud providers
-	_ "k8s.io/kubernetes/pkg/cloudprovider/providers"
+	_ "k8s.io/legacy-cloud-providers/install"
 )
 
 // ProbeVolumePlugins collects all volume plugins into an easy to use list.
