--- pkg/cloudprovider/BUILD
@@ -1,28 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-load(
-    "@io_bazel_rules_go//go:def.bzl",
-    "go_library",
-)
-
-go_library(
-    name = "go_default_library",
-    srcs = ["doc.go"],
-    importpath = "k8s.io/kubernetes/pkg/cloudprovider",
-)
-
-filegroup(
-    name = "package-srcs",
-    srcs = glob(["**"]),
-    tags = ["automanaged"],
-    visibility = ["//visibility:private"],
-)
-
-filegroup(
-    name = "all-srcs",
-    srcs = [
-        ":package-srcs",
-        "//pkg/cloudprovider/providers:all-srcs",
-    ],
-    tags = ["automanaged"],
-)
