--- pkg/cloudprovider/providers/.import-restrictions
@@ -1,5 +1,0 @@
-rules:
-  # prevent any k8s.io/kubernetes imports outside of this package
-  - selectorRegexp: k8s[.]io/kubernetes
-    allowedPrefixes:
-      - k8s.io/kubernetes/pkg/cloudprovider/providers
