+++ staging/src/k8s.io/legacy-cloud-providers/BUILD
@@ -12,6 +12,7 @@ filegroup(
         "//staging/src/k8s.io/legacy-cloud-providers/aws:all-srcs",
         "//staging/src/k8s.io/legacy-cloud-providers/azure:all-srcs",
         "//staging/src/k8s.io/legacy-cloud-providers/gce:all-srcs",
+        "//staging/src/k8s.io/legacy-cloud-providers/install:all-srcs",
         "//staging/src/k8s.io/legacy-cloud-providers/openstack:all-srcs",
         "//staging/src/k8s.io/legacy-cloud-providers/vsphere:all-srcs",
     ],
